feat: add metadata title to address, order-options and order page
